+++ packages/vsce/src/commands/index.ts
@@ -28,12 +28,14 @@ import { getOpenLocalFileCommand } from "./openLocalFileCommand";
 import { getPhaseInCommand } from "./phaseInCommand";
 import { getPurgeTaskCommand } from "./purgeTaskCommand";
 import { getRefreshCommand } from "./refreshCommand";
-import { getResourceInspectorCommand } from "./resourceInspectorViewCommand";
+import { getInspectTreeResourceCommand } from "./inspectTreeResourceCommand";
 import * as showAttributesCommands from "./showAttributesCommand";
+import { showLibraryCommand } from "./showLibraryCommand";
 import * as showLogsCommands from "./showLogsCommand";
 import { getShowRegionSITParametersCommand } from "./showParameterCommand";
 import { viewMoreCommand } from "./viewMoreCommand";
 import { setFocusRegionCommand } from "./setFocusRegionCommand";
+import { getInspectResourceCommand } from "./inspectResourceCommand";
 
 export const getCommands = (treeDataProv: CICSTree, treeview: TreeView<any>, context: ExtensionContext) => {
   return [
@@ -57,6 +59,8 @@ export const getCommands = (treeDataProv: CICSTree, treeview: TreeView<any>, con
 
     getPurgeTaskCommand(treeDataProv, treeview),
 
+    getInspectResourceCommand(context, treeview),
+
     showLogsCommands.getShowRegionLogs(treeview),
     showAttributesCommands.getShowResourceAttributesCommand(treeview),
     showAttributesCommands.getShowRegionAttributes(),
@@ -74,8 +78,9 @@ export const getCommands = (treeDataProv: CICSTree, treeview: TreeView<any>, con
     getInquireTransactionCommand(treeDataProv, treeview),
     getInquireProgramCommand(treeDataProv, treeview),
 
-    getResourceInspectorCommand(context, treeview),
+    getInspectTreeResourceCommand(context, treeview),
     getToggleResourceSettingCommand(),
     setFocusRegionCommand(),
+    showLibraryCommand(treeDataProv, treeview),
   ];
 };