+++ packages/vsce/src/webviews/resource-inspector-panel/ResourceInspector.tsx
@@ -9,15 +9,7 @@
  *
  */
 
-import {
-  VscodeTable,
-  VscodeTableBody,
-  VscodeTableCell,
-  VscodeTableHeader,
-  VscodeTableHeaderCell,
-  VscodeTableRow,
-  VscodeTextfield,
-} from "@vscode-elements/react-elements";
+import { VscodeTextfield } from "@vscode-elements/react-elements";
 
 import * as React from "react";
 import * as vscode from "../common/vscode";
@@ -40,6 +32,25 @@ const ResourceInspector = () => {
       setResourceInfo(event.data.data);
     };
     vscode.addVscMessageListener(listener);
+    const handleScroll = () => {
+      const headerElement1 = document.getElementById("table-header-1");
+      const headerHeight1 = headerElement1.offsetHeight;
+      const headerElement2 = document.getElementById("table-header-2");
+      // Adjust the top position of the second header based on the first header's height
+      headerElement2.style.top = headerHeight1 - 1 + "px";
+    };
+    vscode.addScrollerListener(handleScroll);
+
+    const handleResize = () => {
+      const headerElement1 = document.getElementById("table-header-1");
+      const headerElement2 = document.getElementById("table-header-2");
+      // Adjust the width of both table headers on resize with a offset margin to maintain header alingment
+      if (headerElement1.style.width != getComputedStyle(headerElement2).width) {
+        headerElement1.style.width = Number(getComputedStyle(headerElement2).width.replace("px", "")) - 10 + "px";
+      }
+    };
+    vscode.addResizeListener(handleResize);
+
     vscode.postVscMessage({ command: "init" });
 
     return () => {
@@ -48,67 +59,62 @@ const ResourceInspector = () => {
   }, []);
 
   return (
-    resourceInfo && (
-      <div className="maindiv">
-        <VscodeTable>
-          <VscodeTableHeader>
-            <VscodeTableRow>
-              <VscodeTableHeaderCell className="header-cell-1">
-                <div className="div-display-1">{resourceInfo.name}</div>
-                <div className="div-display-1 div-display-2">
-                  {resourceInfo.resourceName}: {resourceInfo.resource.status || resourceInfo.resource.enablestatus}
-                </div>
-              </VscodeTableHeaderCell>
-            </VscodeTableRow>
-          </VscodeTableHeader>
-          {resourceInfo.highlights.length > 0 && (
-            <VscodeTableBody>
-              <VscodeTableCell className="padding-left-20">
-                {resourceInfo.highlights.map((highlight) => (
-                  <p className="line">
-                    {highlight.key}: {highlight.value}
-                  </p>
-                ))}
-              </VscodeTableCell>
-            </VscodeTableBody>
-          )}
-        </VscodeTable>
-
-        <VscodeTable zebra={true} columns={["30%", "70%"]}>
-          <VscodeTableHeader>
-            <VscodeTableRow>
-              <VscodeTableHeaderCell className="div-display-1 padding-right-10">Attribute</VscodeTableHeaderCell>
-              <VscodeTableHeaderCell className="padding-right-10">
-                <div>
-                  <div className="div-display-1">Value</div>
-                  <VscodeTextfield
-                    type="text"
-                    placeholder="Keyword search..."
-                    onInput={(e: { target: HTMLInputElement; }) => setSearch(e.target.value)}
-                    value={search}
-                    className="search-style div-display-1"
-                  ></VscodeTextfield>
-                </div>
-              </VscodeTableHeaderCell>
-            </VscodeTableRow>
-          </VscodeTableHeader>
-          <VscodeTableBody>
-            {Object.entries(resourceInfo.resource)
-              .filter(([key, value]) => !key.startsWith("_"))
-              .filter(
-                ([key, value]) =>
-                  key.toLowerCase().trim().includes(search.toLowerCase().trim()) || value.toLowerCase().trim().includes(search.toLowerCase().trim())
-              )
-              .map(([key, value]) => (
-                <VscodeTableRow>
-                  <VscodeTableCell className="padding-left-20">{key}</VscodeTableCell>
-                  <VscodeTableCell className="padding-right-75">{value}</VscodeTableCell>
-                </VscodeTableRow>
+    <div className="maindiv" data-vscode-context='{"webviewSection": "main", "mouseCount": 4}'>
+      <table id="table-1" className="border-collapse">
+        <thead id="table-header-1" className="table-header1">
+          <th id="th-1" className="header-cell-1 padding-left-10">
+            <div className="div-display-1">{resourceInfo?.name ?? "..."}</div>
+            <div className="div-display-1 div-display-2">
+              {resourceInfo?.resourceName ?? "..."}: {resourceInfo ? (resourceInfo?.resource.status || resourceInfo.resource.enablestatus) : "..."}
+            </div>
+          </th>
+        </thead>
+        <tbody className="padding-left-10 padding-top-20">
+          {resourceInfo?.highlights.length > 0 && (
+            <tr>
+              <p className="padding-top-10"></p>
+              {resourceInfo.highlights.map((highlight) => (
+                <p className="line padding-left-20">
+                  {highlight.key}: {highlight.value}
+                </p>
               ))}
-          </VscodeTableBody>
-        </VscodeTable>
-      </div>
-    )
+            </tr>
+          )}
+        </tbody>
+      </table>
+
+      <table className="border-collapse">
+        <thead id="table-header-2" className="thead-2 vertical-align-sub">
+          <th className="div-display-1 th-2">Attribute</th>
+          <th className="padding-right-10 th-3">
+            <div>
+              <div className="div-display-1 vertical-align-sub">Value</div>
+              <VscodeTextfield
+                type="text"
+                placeholder="Keyword search..."
+                onInput={(e: { target: HTMLInputElement; }) => setSearch(e.target.value)}
+                value={search}
+                className="search-style div-display-1"
+              ></VscodeTextfield>
+            </div>
+          </th>
+        </thead>
+        <tbody>
+          {resourceInfo && Object.entries(resourceInfo.resource)
+            .filter(([key, value]) => !key.startsWith("_"))
+            .filter(
+              ([key, value]) =>
+                key.toLowerCase().trim().includes(search.toLowerCase().trim()) || value.toLowerCase().trim().includes(search.toLowerCase().trim())
+            )
+            .map(([key, value]) => (
+              <tr key={key}>
+                <td className="padding-left-27 width-30">{key}</td>
+                <td className="padding-right-75 width-70">{value}</td>
+              </tr>
+            ))}
+        </tbody>
+      </table>
+    </div>
   );
 };